Changed from host/client to y/n for easier typing for the user.
Added validation for choice variable in main so only either y/n can be typed in the field.

Changed startAsCLient to now wait if there is no server active. This is done via adding a try/catch statement in the method, which endlessly tries to connect to a server via socket. If IOException is present, the catch block will handle it and it will loop to the top again. If connection is present, the server will send a first message READY, which the client will check if it has been sent. If it has, the game can start. If not, the client will wait. Potentially need to change from a busy-loop. but its working in the meanwhile.