Changed the host variable from localhost to the local IP. This allows for a LAN connection, two machines on the same network instead of one machine.
Added functionality: username. Users now have to type in a username before connecting. This was done by adding a username field to the ConnectionManager constructor. This will contain the username of the user. Changes to startAsServer() and startAsClient() were needed, due to them now sending their usernames to each other. setUpStreams() is now called inside those two methods, ensuring that the username is sent after sockets, in and out are intilized. Host and client are not sending the usernames at the same time, due to the risk of deadlock. Client sends first, and then the server sents.

Getter was added for opponents username inside ConnectionManager.